feat: AI 봇 목록 API 및 SwaggerDoc 설정 완료
- SpringDoc OpenAPI 2.6.0  2.8.0 업그레이드 (Spring Boot 3.5.4 호환)
- AI 봇 목록 조회 API 구현 (GET /api/ai-bots)
- Entity 타입 정렬 (Long  Integer, DB 스키마와 일치)
- 환경별 설정 분리 (로컬/Docker/테스트)
- OpenAPI 설정 추가 (Swagger UI 서버 정보 포함)
- EC2 MySQL 직접 연결 설정
- 더미 AI 봇 데이터 연동 테스트 완료

+++ BE/src/main/java/com/example/LAGO/repository/MockTradeRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface MockTradeRepository extends JpaRepository<MockTrade, Long> {
-    Long countByAccountId(Long accountId);
+    Long countByAccountId(Integer accountId);
 
     @Query("SELECT AVG(mt.price * mt.quantity) FROM MockTrade mt WHERE mt.account.accountId = :accountId")
-    Double findAvgTradeValue(@Param("accountId") Long accountId);
+    Double findAvgTradeValue(@Param("accountId") Integer accountId);
 }
